Set connection in add and remove feedback

+++ app/src/main/java/com/example/igor/projetopoo/activity/product/ProductModel.java
@@ -1,6 +1,7 @@
 package com.example.igor.projetopoo.activity.product;
 
 import android.support.constraint.ConstraintLayout;
+import android.util.Log;
 import android.widget.ScrollView;
 
 import com.example.igor.projetopoo.R;
@@ -74,15 +75,42 @@ public class ProductModel implements ProductMVP.ModelOps {
     }
 
     @Override
-    public void insertFeedback(Feedback feedback) throws DatabaseException {
+    public void insertFeedback(Feedback feedback) throws ConnectionException, DatabaseException {
+        final ConstraintLayout layout = activity.findViewById(R.id.container_product);
+
+        activity.runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                layout.removeAllViews();
+            }
+        });
+
+        if (!ParentActivity.checkConnection(activity)) {
+            throw new ConnectionException(activity, layout);
+        }
+
         FirebaseFirestore firestore = database.getFirestore();
         task = database.addDocument(firestore.collection("feedbacks"),feedback);
         if (!task.isSuccessful()) throw new DatabaseException(activity);
+        Log.i("TAG", "EAE");
         reqPresenterOps.onFeedbackInserted();
     }
 
     @Override
-    public void deleteFeedback() throws DatabaseException {
+    public void deleteFeedback() throws ConnectionException, DatabaseException {
+        final ConstraintLayout layout = activity.findViewById(R.id.container_product);
+
+        activity.runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                layout.removeAllViews();
+            }
+        });
+
+        if (!ParentActivity.checkConnection(activity)) {
+            throw new ConnectionException(activity, layout);
+        }
+
         database.deleteDocument(task.getResult());
         if (!task.isSuccessful()) throw new DatabaseException(activity);
         reqPresenterOps.onFeedbackDeleted();
